Fix profile-back-nav bug, add Elo chart to public profiles, style tables
Three items from BACKEND_REWORK.MD/FRONTEND_REWORK.MD:

- Fix a real reported navigation bug: Home -> Recent Games -> click a
  game -> click a player -> Back landed on Leaderboard instead of Home.
  The game detail modal now captures which real screen was showing
  underneath it the instant it opens (Home, My Games, Account, or
  another Player Profile), and passes that through as the Back target
  instead of always assuming Leaderboard.

- Public profiles now show the Elo chart too (previously Account-only --
  not a bug, just an initial scope cut, now closed). Pulled the whole
  ApexCharts implementation out of account.js into a new shared
  ui/eloChart.js (createEloChartController), so both screens draw from
  one implementation instead of a second copy -- Account's own chart is
  unchanged behavior, just reusing the extracted code. Shown for bots
  too, since they're real rated participants with a genuine Elo trend,
  not a placeholder. Added a relevance guard so navigating from one
  profile to another before a slow chart fetch resolves can't paint
  stale data into the wrong screen -- a real risk here that Account
  never had (no login/logout cycle slowing that down).

- Styled the shared .leaderboard-table (used by both the Leaderboard
  screen and the game detail modal's "Game Results" table, flagged as
  bland/monotone in both places): alternating row tint, a clickable
  name now colored + underlined at rest instead of only on hover, and a
  slightly heavier header divider.

Verified via Playwright (13 checks): Home -> game -> bot profile -> Back
returns to Home; a bot with no known-difficulty backing identity renders
as plain text, not a dead link; a bot's public profile shows a real,
working Elo chart with its own range toggle; both tables show visible
zebra striping and colored name-links at rest. Full backend suite green
(526 passed, unaffected -- frontend-only change).
